add no tuner message and dataclass
