+++ nni/experiment/__init__.py
@@ -2,4 +2,5 @@
 # Licensed under the MIT license.
 
 from .config import *
-from .experiment import Experiment, TrialJob, TrialResult, TrialMetricData, TrialHyperParameters
+from .experiment import Experiment
+from .data import *
