+++ nni/experiment/experiment.py
@@ -1,5 +1,4 @@
 import atexit
-import json
 import logging
 from pathlib import Path
 import socket
@@ -16,6 +15,7 @@ from nni.runtime.msg_dispatcher import MsgDispatcher
 from nni.tuner import Tuner
 
 from .config import ExperimentConfig
+from .data import TrialJob, TrialMetricData, TrialResult
 from . import launcher
 from . import management
 from .pipe import Pipe
@@ -106,7 +106,8 @@ class Experiment:
                 self.config = ExperimentConfig(training_service)
             else:
                 self.config = config
-
+        else:
+            _logger.warning('Tuner not set, wait for connect...')
 
     def start(self, port: int = 8080, debug: bool = False) -> None:
         """
@@ -153,10 +154,8 @@ class Experiment:
         msg = 'Web UI URLs: ' + colorama.Fore.CYAN + ' '.join(ips) + colorama.Style.RESET_ALL
         _logger.info(msg)
 
-
     def _create_dispatcher(self):  # overrided by retiarii, temporary solution
         return MsgDispatcher(self.tuner, None)
-
 
     def stop(self) -> None:
         """
@@ -183,7 +182,6 @@ class Experiment:
         self._dispatcher_thread = None
         _logger.info('Experiment stopped')
 
-
     def run(self, port: int = 8080, debug: bool = False) -> bool:
         """
         Run the experiment.
@@ -206,7 +204,6 @@ class Experiment:
         finally:
             self.stop()
 
-
     def connect_experiment(self, port: int):
         """
         Connect to an existing experiment.
@@ -219,18 +216,15 @@ class Experiment:
         self.port = port
         self.get_status()
 
-
     def _experiment_rest_get(self, port: int, api: str) -> Any:
         if self.port is None:
             raise RuntimeError('Experiment is not running')
         return rest.get(self.port, api)
 
-
     def _experiment_rest_put(self, port: int, api: str, data: Any):
         if self.port is None:
             raise RuntimeError('Experiment is not running')
         rest.put(self.port, api, data)
-
 
     def get_status(self) -> str:
         """
@@ -243,7 +237,6 @@ class Experiment:
         """
         resp = self._experiment_rest_get(self.port, '/check-status')
         return resp['status']
-
 
     def get_trial_job(self, trial_job_id: str):
         """
@@ -260,8 +253,7 @@ class Experiment:
             A `TrialJob` instance corresponding to `trial_job_id`.
         """
         resp = self._experiment_rest_get(self.port, '/trial-jobs/{}'.format(trial_job_id))
-        return TrialJob(resp)
-
+        return TrialJob(**resp)
 
     def list_trial_jobs(self):
         """
@@ -273,8 +265,7 @@ class Experiment:
             List of `TrialJob`.
         """
         resp = self._experiment_rest_get(self.port, '/trial-jobs')
-        return [TrialJob(trial_job) for trial_job in resp]
-
+        return [TrialJob(**trial_job) for trial_job in resp]
 
     def get_job_statistics(self):
         """
@@ -287,7 +278,6 @@ class Experiment:
         """
         resp = self._experiment_rest_get(self.port, '/job-statistics')
         return resp
-
 
     def get_job_metrics(self, trial_job_id=None):
         """
@@ -309,11 +299,10 @@ class Experiment:
         for metric in resp:
             trial_id = metric["trialJobId"]
             if trial_id not in metric_dict:
-                metric_dict[trial_id] = [TrialMetricData(metric)]
+                metric_dict[trial_id] = [TrialMetricData(**metric)]
             else:
-                metric_dict[trial_id].append(TrialMetricData(metric))
+                metric_dict[trial_id].append(TrialMetricData(**metric))
         return metric_dict
-
 
     def get_experiment_profile(self):
         """
@@ -327,7 +316,6 @@ class Experiment:
         resp = self._experiment_rest_get(self.port, '/experiment')
         return resp
 
-
     def export_data(self):
         """
         Return exported information for all trial jobs.
@@ -338,8 +326,7 @@ class Experiment:
             List of `TrialResult`.
         """
         resp = self._experiment_rest_get(self.port, '/export-data')
-        return [TrialResult(trial_result) for trial_result in resp]
-
+        return [TrialResult(**trial_result) for trial_result in resp]
 
     def _get_query_type(self, key: str):
         if key == 'trial_concurrency':
@@ -350,7 +337,6 @@ class Experiment:
             return '?update_type=SEARCH_SPACE'
         if key == 'max_trial_number':
             return '?update_type=MAX_TRIAL_NUM'
-
 
     def _update_experiment_profile(self, key: str, value: Any):
         """
@@ -368,7 +354,6 @@ class Experiment:
         experiment_profile['params'][key] = value
         self._experiment_rest_put(self.port, api, experiment_profile)
 
-
     def update_trial_concurrency(self, value: int):
         """
         Update an experiment's trial_concurrency
@@ -379,7 +364,6 @@ class Experiment:
             New trial_concurrency value.
         """
         self._update_experiment_profile('trial_concurrency', value)
-
 
     def update_max_experiment_duration(self, value: str):
         """
@@ -393,7 +377,6 @@ class Experiment:
         """
         self._update_experiment_profile('max_experiment_duration', value)
 
-
     def update_search_space(self, value: dict):
         """
         Update the experiment's search_space.
@@ -406,7 +389,6 @@ class Experiment:
         """
         self._update_experiment_profile('search_space', value)
 
-
     def update_max_trial_number(self, value):
         """
         Update an experiment's max_trial_number
@@ -417,159 +399,3 @@ class Experiment:
             New max_trial_number value.
         """
         self._update_experiment_profile('max_trial_number', value)
-
-
-class TrialResult:
-    """
-    TrialResult stores the result information of a trial job.
-
-    Parameters
-    ----------
-    json_obj: dict
-        Json object that stores the result information.
-
-    Attributes
-    ----------
-    parameter: dict
-        Hyper parameters for this trial.
-    value: serializable object, usually a number, or a dict with key "default" and other extra keys
-        Final result.
-    trialJobId: str
-        Trial job id.
-    """
-    def __init__(self, json_obj):
-        self.parameter = None
-        self.value = None
-        self.trialJobId = None
-        for key in json_obj.keys():
-            setattr(self, key, json_obj[key])
-        self.value = json.loads(self.value)
-
-    def __repr__(self):
-        return "TrialResult(parameter: {} value: {} trialJobId: {})".format(self.parameter, self.value, self.trialJobId)
-
-
-class TrialMetricData:
-    """
-    TrialMetricData stores the metric data of a trial job.
-    A trial job may have both intermediate metric and final metric.
-
-    Parameters
-    ----------
-    json_obj: dict
-        Json object that stores the metric data.
-
-    Attributes
-    ----------
-    timestamp: int
-        Time stamp.
-    trialJobId: str
-        Trial job id.
-    parameterId: int
-        Parameter id.
-    type: str
-        Metric type, `PERIODICAL` for intermediate result and `FINAL` for final result.
-    sequence: int
-        Sequence number in this trial.
-    data: serializable object, usually a number, or a dict with key "default" and other extra keys
-        Metric data.
-    """
-    def __init__(self, json_obj):
-        self.timestamp = None
-        self.trialJobId = None
-        self.parameterId = None
-        self.type = None
-        self.sequence = None
-        self.data = None
-        for key in json_obj.keys():
-            setattr(self, key, json_obj[key])
-        self.data = json.loads(json.loads(self.data))
-
-    def __repr__(self):
-        return "TrialMetricData(timestamp: {} trialJobId: {} parameterId: {} type: {} sequence: {} data: {})" \
-            .format(self.timestamp, self.trialJobId, self.parameterId, self.type, self.sequence, self.data)
-
-
-class TrialHyperParameters:
-    """
-    TrialHyperParameters stores the hyper parameters of a trial job.
-
-    Parameters
-    ----------
-    json_obj: dict
-        Json object that stores the hyper parameters.
-
-    Attributes
-    ----------
-    parameter_id: int
-        Parameter id.
-    parameter_source: str
-        Parameter source.
-    parameters: dict
-        Hyper parameters.
-    parameter_index: int
-        Parameter index.
-    """
-    def __init__(self, json_obj):
-        self.parameter_id = None
-        self.parameter_source = None
-        self.parameters = None
-        self.parameter_index = None
-        for key in json_obj.keys():
-            if hasattr(self, key):
-                setattr(self, key, json_obj[key])
-
-    def __repr__(self):
-        return "TrialHyperParameters(parameter_id: {} parameter_source: {} parameters: {} parameter_index: {})" \
-            .format(self.parameter_id, self.parameter_source, self.parameters, self.parameter_index)
-
-
-class TrialJob:
-    """
-    TrialJob stores the information of a trial job.
-
-    Parameters
-    ----------
-    json_obj: dict
-        json object that stores the hyper parameters
-
-    Attributes
-    ----------
-    trialJobId: str
-        Trial job id.
-    status: str
-        Job status.
-    hyperParameters: list of `nni.experiment.TrialHyperParameters`
-        See `nni.experiment.TrialHyperParameters`.
-    logPath: str
-        Log path.
-    startTime: int
-        Job start time (timestamp).
-    endTime: int
-        Job end time (timestamp).
-    finalMetricData: list of `nni.experiment.TrialMetricData`
-        See `nni.experiment.TrialMetricData`.
-    parameter_index: int
-        Parameter index.
-    """
-    def __init__(self, json_obj):
-        self.trialJobId = None
-        self.status = None
-        self.hyperParameters = None
-        self.logPath = None
-        self.startTime = None
-        self.endTime = None
-        self.finalMetricData = None
-        self.stderrPath = None
-        for key in json_obj.keys():
-            setattr(self, key, json_obj[key])
-        if self.hyperParameters:
-            self.hyperParameters = [TrialHyperParameters(json.loads(e)) for e in self.hyperParameters]
-        if self.finalMetricData:
-            self.finalMetricData = [TrialMetricData(e) for e in self.finalMetricData]
-
-    def __repr__(self):
-        return ("TrialJob(trialJobId: {} status: {} hyperParameters: {} logPath: {} startTime: {} "
-                "endTime: {} finalMetricData: {} stderrPath: {})") \
-                .format(self.trialJobId, self.status, self.hyperParameters, self.logPath,
-                        self.startTime, self.endTime, self.finalMetricData, self.stderrPath)
